Greatly rewrite large parts of 'training' - especially getting started
The guide was overcomplex, lacked grouping (storage in 3 places, dns in
4) did the same operation multiple times. Did operations not strictly
necessary. etc.

This should greatly simplify the steps to get the cluster up. There is
obviously serious work needed in the later sections and we just straight
into scheduling constraints before we even talk about what a pod is, but
I didn't get that far.